Fast-CRC: Fast CRC cleanup.
I tried unifying the common aspects in a kernel like the 4-bit tables.
This proved to be to cumbersome -- at least on this attempt.

Instead, I swapped out the uint32_t return types where possible
for the natural sized CRC types.

This saved the need to mask results (in code),
but I guess the compiler will need to mask registers anyway.

WIP..