Fast-CRC: Refactor CRC 4-bit Table-Kernel - Stage 1
The current CRC kernels abuse #include in the hope of inlining everything:
- Polynomial headers declare their kernel requirements,
- which are included by the various algorithms,
- which are included in the application defined "crc_algorithms.h",
- which is included by "crc.h" via build defined include path,
- which is included by "crc.c" and processed to determined the actual kernel requirements,
- which then includes one or more named kernels,
- which each define the same boiler-plate calc_crc{x,8,16,32,64} functions,
- which take a function pointer to a type-neutering polynomial table wrapper

Ahhh!

They work, but they miss the main requirement, namely providing small + speedy CRC implementations.
I argued myself into a corner about the function inlining...
I believed that the compiler would figure out that everything could be inlined,
given that everything was #included into the same compilation unit.

Was I wrong?
I'm not sure, but I'm not going to bother finding out.

I'm rewriting the kernel generation mechanism to use macros.
It's verbose and ugly, but easy to understand and maintain.

There's a lot of repetition,
but it's easy to interpolate the changes between forward and reverse kernels,
and different CRC widths.

I've elected to use the uint_fast{8,16,32,64}_t family of types,
which are defined in C99 and later.

On 32-bit architectures, the 16 and 32 variants are aliases for uint32_t
This also holds for Windows platforms \(both ILP32 and LLP64\).

On 64-bit linux platforms, the 16 and 32 variants are aliases for uint64_t.

All platforms alias fast8 to uint8_t AFAICT.

What this means is that -- in practice -- these types do the right thing.
The _impl kernel methods use these faster integers for processing.
And the interfacing methods simply mask the result to the required CRC width.