Fast-CRC: Bug fix - Fast Algorithms calling the Table-Kernel.
All of the Fast CRC Algorithms were inadvertently calling the Table Kernel via the make_crcxx_implementation macro...

I removed this macro call and replaced with hand-edited interfaces that call the actual Fast CRC functions.
This revealed a data_len overflow bug in three of the algorithms, which I fixed.

All tests pass with changes, but it highlights the need to more debuggable tests.
I'll address this shortly.