Fast-CRC: Updated tools/make_polynomial_db to process Hamming Distance Survery Data.
The functionality is currently gated by the PROCESS_SURVEY_DATA variable.
This is because the survey data contains over 30,000 entries.

I think it'd be a bad idea to have 30,000+ JSON5 files in the checkout.
So I'll probably update the process to create a tar ball of the profiles and the db.

This should suit HTTP(S) -- need to check...